requirements.txt
streamlit==1.28.1
pandas==2.1.3
plotly==5.18.0
google-generativeai>=0.6.0
reportlab==4.0.7
openpyxl==3.1.5
python-dotenv==1.0.0
